reranker/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]